+++ examples/train_causal_lm.py
@@ -19,7 +19,7 @@ from peft import LoraConfig, get_peft_model
 from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
 from opaque.clipping import clipped_grad
-from opaque.noise import add_gaussian_noise
+from opaque.noise import gaussian_noise
 from opaque.utils import make_functional
 
 
@@ -373,7 +373,8 @@ def main():
 
             # Add Gaussian noise
             stddev = args.noise_multiplier * current_clip_norm
-            noisy_grads = add_gaussian_noise(grads_tuple, stddev=stddev, generator=rng)
+            noise_fn, noise_state = gaussian_noise(stddev=stddev, generator=rng)
+            noisy_grads, _ = noise_fn(grads_tuple, noise_state)
 
             # Optimizer step
             if args.use_adaptive_clipping: